# docs/web-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TalkrAI web API",
    "version": "1.0.0",
    "description": "Same-origin Workers proxy to Rust. UUIDs are opaque; owner-scoped resources return 404 across owners. Story persona/version mutations conflict with in-flight turns. No provider tokens in responses. OAuth callback errors redirect to a fixed login page."
  },
  "paths": {
    "/api/public/scenes": {
      "get": {
        "operationId": "get__api_public_scenes",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenePage"
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 expired; 402 credits; 403 consent/suspension/CSRF; 404 resource; 409 conflict; 429 rate; 502 upstream; 503 admissions closed. Private responses are no-store.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/public/scenes/{id}": {
      "get": {
        "operationId": "get__api_public_scenes_id",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scene"
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 expired; 402 credits; 403 consent/suspension/CSRF; 404 resource; 409 conflict; 429 rate; 502 upstream; 503 admissions closed. Private responses are no-store.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/web/auth/{provider}/start": {
      "get": {
        "operationId": "get__api_web_auth_provider_start",
        "responses": {
          "303": {
            "description": "Success"
          },
          "default": {
            "description": "400 validation; 401 expired; 402 credits; 403 consent/suspension/CSRF; 404 resource; 409 conflict; 429 rate; 502 upstream; 503 admissions closed. Private responses are no-store.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "google",
                "line"
              ]
            }
          },
          {
            "name": "return_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/web/auth/{provider}/callback": {
      "get": {
        "operationId": "get__api_web_auth_provider_callback",
        "responses": {
          "303": {
            "description": "Success"
          },
          "default": {
            "description": "400 validation; 401 expired; 402 credits; 403 consent/suspension/CSRF; 404 resource; 409 conflict; 429 rate; 502 upstream; 503 admissions closed. Private responses are no-store.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "google",
                "line"
              ]
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/web/me": {
      "get": {
        "operationId": "get__api_web_me",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 expired; 402 credits; 403 consent/suspension/CSRF; 404 resource; 409 conflict; 429 rate; 502 upstream; 503 admissions closed. Private responses are no-store.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "webSession": []
          }
        ]
      }
    },
    "/api/web/credits": {
      "get": {
        "operationId": "get__api_web_credits",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Credits"
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 expired; 402 credits; 403 consent/suspension/CSRF; 404 resource; 409 conflict; 429 rate; 502 upstream; 503 admissions closed. Private responses are no-store.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "webSession": []
          }
        ]
      }
    },
    "/api/web/credit-transactions": {
      "get": {
        "operationId": "get__api_web_credit-transactions",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionPage"
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 expired; 402 credits; 403 consent/suspension/CSRF; 404 resource; 409 conflict; 429 rate; 502 upstream; 503 admissions closed. Private responses are no-store.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "webSession": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/web/sessions": {
      "get": {
        "operationId": "get__api_web_sessions",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryPage"
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 expired; 402 credits; 403 consent/suspension/CSRF; 404 resource; 409 conflict; 429 rate; 502 upstream; 503 admissions closed. Private responses are no-store.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "webSession": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      },
      "post": {
        "operationId": "post__api_web_sessions",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Story"
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 expired; 402 credits; 403 consent/suspension/CSRF; 404 resource; 409 conflict; 429 rate; 502 upstream; 503 admissions closed. Private responses are no-store.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "webSession": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scene_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "persona": {
                    "$ref": "#/components/schemas/Persona"
                  }
                },
                "required": [
                  "scene_id",
                  "persona"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ]
      }
    },
    "/api/web/sessions/{id}": {
      "get": {
        "operationId": "get__api_web_sessions_id",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Story"
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 expired; 402 credits; 403 consent/suspension/CSRF; 404 resource; 409 conflict; 429 rate; 502 upstream; 503 admissions closed. Private responses are no-store.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "webSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/web/sessions/{id}/messages": {
      "get": {
        "operationId": "get__api_web_sessions_id_messages",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagePage"
                }
              }
            }
          },
          "default": {
            "description": "400 validation; 401 expired; 402 credits; 403 consent/suspension/CSRF; 404 resource; 409 conflict; 429 rate; 502 upstream; 503 admissions closed. Private responses are no-store.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "webSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          },
          {
            "name": "before",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "webSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-talkrai"
      }
    },
    "schemas": {
...
      "Me": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "linked_providers": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "google",
                "line"
              ]
            }
          },
          "terms_accepted": {
            "type": "boolean"
          },
          "terms_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "required_terms_version": {
            "type": "string"
          },
          "account_status": {
            "type": "string",
            "enum": [
              "active",
              "suspended"
            ]
          },
          "csrf_token": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "linked_providers",
          "terms_accepted",
          "terms_version",
          "required_terms_version",
          "account_status",
          "csrf_token"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ]
      },
...
      "ScenePage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Scene"
            }
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "items",
          "next_cursor"
        ]
      },
...
    }
  }
}